requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
azure-core==1.30.0
azure-common==1.1.28
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
